Enhancement and cleanup of video processing
Added some additional register definitions.
Added initialisation of all key registers, following the code which
was in the original Techwell bttv patches.
Changed norm data to be table-based, with a table for each of the
norms that we handle.
Fixed serious error in the risc code generation.  With this change,
program can now display a picture with xawtv (but apparently the
norm constants are not correct, and need further work, so there is still
a lot of room for improvement).