Enhance for v4l2 changes in kernels 2.6.35 and 2.6.36
For tw68-video.c, changed second parameter in calls to v4l2_prio_check
and v4l2_prio_close to be enum instead of pointer to enum.
For tw68-core.c, changed first parameter in call to videobuf_dma_unmap
to be pointer to device structure instead of pointer to video buff.
Surrounded by change by #if for kernel version so it should work for
older kernels as well as more recent ones.